cuda build break, and sequence op int vs size mismatch

+++ orttraining/orttraining/training_ops/cpu/optimizer/clip_grad_norm/clip_grad_norm.cc
@@ -28,7 +28,7 @@ template <typename T>
 void ClipGradNorm(T total_norm, T max_norm, TensorSeq& gradients) {
   const T clip_coefficient = std::min(max_norm / (total_norm + static_cast<T>(Epsilon)), static_cast<T>(1.0f));
 
-  for (int i = 0; i < gradients.Size(); i++)
+  for (size_t i = 0; i < gradients.Size(); i++)
   {
     const auto& grad = gradients.Get(i);
     auto& tensor = const_cast<Tensor&>(grad);
@@ -46,7 +46,7 @@ Status PopulateOutput(OpKernelContext* ctx, const TensorSeq* gradients, TensorSe
 
   clipped_gradients->SetType(gradients->DataType());
   clipped_gradients->SetElements({});
-  for (int i = 0; i < gradients->Size(); i++)
+  for (size_t i = 0; i < gradients->Size(); i++)
   {
     const auto& grad = gradients->Get(i);
     Tensor target_tensor(grad.DataType(), grad.Shape(), alloc);
